requirement.txt
streamlit
opencv-python
pytesseract
numpy
pillow
nltk
spacy
transformers
torch